client/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
insightface>=0.7.0
onnxruntime>=1.15.0
requests>=2.28.0
PyYAML>=6.0
supabase>=2.0.0
